requirements.txt
pyqt
matplotlib
